Update packaging and CI
