--- .fussyfox.yml
@@ -1,4 +1,0 @@
-- bandit
-- flake8
-- isort
-- pydocstyle
