+++ .github/workflows/ci.yml
@@ -8,6 +8,40 @@ on:
 
 jobs:
 
+  dist:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: "3.x"
+      - run: python -m pip install --upgrade pip build wheel twine
+      - run: python -m build --sdist --wheel
+      - run: python -m twine check dist/*
+      - uses: actions/upload-artifact@v3
+        with:
+          path: dist/*
+
+  lint:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        lint-command:
+          - bandit -r . -x ./tests
+          - black --check --diff .
+          - flake8 .
+          - isort --check-only --diff .
+          - pydocstyle .
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: "3.x"
+          cache: 'pip'
+          cache-dependency-path: 'linter-requirements.txt'
+      - run: python -m pip install -r linter-requirements.txt
+      - run: ${{ matrix.lint-command }}
+
   PyTest:
     runs-on: ubuntu-latest
     strategy:
@@ -22,8 +56,8 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
-    - run: python -m pip install --upgrade pip setuptools codecov
+    - run: python -m pip install --upgrade pip setuptools
     - run: python -m pip install -e .[test]
-    - run: relint **
+    - run: relint --version
     - run: py.test --cov=.
     - uses: codecov/codecov-action@v3
