+++ .github/workflows/release.yml
@@ -12,7 +12,7 @@ jobs:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v4
       with:
-        python-version: "3.10"
+        python-version: "3.x"
     - run: python -m pip install --upgrade pip build wheel twine
     - run: python -m build --sdist --wheel
     - run: python -m twine upload dist/*
