+++ linter-requirements.txt
@@ -1,0 +1,5 @@
+bandit==1.7.4
+black==22.8.0
+flake8==5.0.4
+isort==5.10.1
+pydocstyle[toml]==6.1.1
