+++ pyproject.toml
@@ -1,0 +1,72 @@
+[build-system]
+requires = ["flit_core>=3.2", "flit_scm", "wheel"]
+build-backend = "flit_scm:buildapi"
+
+[project]
+name = "relint"
+authors = [
+  { name = "Johannes Maron", email = "johannes@maron.family" },
+]
+readme = "README.md"
+license = { file = "LICENSE" }
+keywords = ["regex", "linter"]
+dynamic = ["version", "description"]
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+  "Environment :: Console",
+  "Intended Audience :: Developers",
+  "License :: OSI Approved :: MIT License",
+  "Operating System :: OS Independent",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Topic :: Software Development",
+  "Topic :: Software Development :: Quality Assurance",
+  "Topic :: Software Development :: Testing",
+]
+requires-python = ">=3.8"
+dependencies = ["PyYAML"]
+
+[project.optional-dependencies]
+test = [
+  "pytest",
+  "pytest-cov",
+  "pytest-mock",
+]
+
+[project.scripts]
+relint = "relint.__main__:main"
+
+[project.urls]
+Project-URL = "https://github.com/codingjoe/relint"
+Changelog = "https://github.com/codingjoe/relint/releases"
+
+[tool.flit.module]
+name = "relint"
+
+[tool.setuptools_scm]
+write_to = "relint/_version.py"
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "--cov --tb=short -rxs"
+testpaths = ["tests"]
+
+[tool.coverage.run]
+source = ["relint"]
+
+[tool.coverage.report]
+show_missing = true
+
+[tool.isort]
+atomic = true
+line_length = 88
+multi_line_output = 3
+force_grid_wrap = 0
+known_first_party = "relint, tests"
+include_trailing_comma = true
+use_parentheses = true
+default_section = "THIRDPARTY"
+combine_as_imports = true
+
+[tool.pydocstyle]
+add_ignore = "D1"
