+++ relint/__init__.py
@@ -1,0 +1,6 @@
+"""Write your own linting rules using regular expressions."""
+
+from . import _version
+
+__version__ = _version.version
+VERSION = _version.version_tuple
