+++ relint/exceptions.py
@@ -1,0 +1,6 @@
+class RelintError(Exception):
+    pass
+
+
+class ConfigError(ValueError, RelintError):
+    pass
