+++ setup.cfg
@@ -1,58 +1,4 @@
-[metadata]
-name = relint
-author = Johannes Maron
-author_email = johannes@maron.family
-description = Write your own linting rules using regular expressions
-long_description = file: README.rst
-url = https://github.com/codingjoe/relint
-license = MIT
-license_file = LICENSE
-classifier =
-    Development Status :: 5 - Production/Stable
-    Environment :: Console
-    Intended Audience :: Developers
-    License :: OSI Approved :: MIT License
-    Intended Audience :: Information Technology
-    Operating System :: OS Independent
-    Programming Language :: Python
-    Topic :: Software Development :: Quality Assurance
-    Topic :: Software Development :: Testing
-keywords =
-    linter
-    regex
-
-[options]
-install_requires = PyYAML
-setup_requires =
-    setuptools_scm
-py_modules = relint
-
-[options.extras_require]
-test =
-    pytest
-    pytest-cov
-    pytest-mock
-
-[options.entry_points]
-console_scripts =
-    relint = relint:main
-
-[bdist_wheel]
-universal = 1
-
-[aliases]
-test = pytest
-
 [flake8]
-max_line_length = 88
-
-[pydocstyle]
-add_ignore = D1
-
-[isort]
-atomic = true
-line_length = 88
-known_first_party = relint
-include_trailing_comma = True
-default_section=THIRDPARTY
-combine_as_imports = true
+max-line-length=88
+select = C,E,F,W,B,B950
+ignore = E203, E501, W503, E731
