--- setup.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-from setuptools import setup
-
-setup(name='relint', use_scm_version=True)
